# Conflicts:
#	enigmathique/frontend/src/App.jsx

+++ enigmathique/backend/src/controllers/game.controller.js
@@ -27,6 +27,7 @@ function makeid(length) {
     return result;
 }
 
+// to do : on peut simplfier ?
 // Fonction vérifiant si une classe, à partir de son id, appartiant au professeur
 async function isClassBelongsProfessor(idCourse, req) {
 	try {
@@ -47,20 +48,20 @@ async function isClassBelongsProfessor(idCourse, req) {
 	}
 }
 
-// Fonction vérifiant si une classe, à partir de son id, appartiant au professeur
+// Fonction vérifiant si une partie, à partir de son id, appartiant au professeur
 async function isGameBelongsProfessor(idGame, req) {
 	try {
 
-		// Récupère toutes les classes du professeur courant
+		// Récupère la partie souhaité
 		const data = await Game.findOne({ where: { id: idGame} });
 		if(data){
-			// Récupère les id correspondant aux classes du professeur courant
-			const idGame = data.id;
+			// Récupère les id des classes des parties
+			const idGame = data.idCourse;
 
-			// Vérifie que la classe appartient bien au professeur
+			// Vérifie que la classe appartiennent bien au professeur
 			return await isClassBelongsProfessor(idGame, req);
 		} else {
-			throw new Error("L'élève n'existe pas.");
+			throw new Error("La parrtie n'existe pas.");
 		}
 
 	} catch (err) {
@@ -74,17 +75,24 @@ async function isGameBelongsProfessor(idGame, req) {
 /////////////////////////////////////////////////////////////////////////////////
 
 
-// Créer et enregistrer une nouvelle classe au professeurs
+// Créer et enregistrer une nouvelle partie
 exports.create = async (req, res) => {
 
 	// Valider la requête
 	if (!req.body.idCourse) {
 		return res.status(400).json({
-			message: "Il manque des informations pour créer la classe."
+			message: "Il manque des informations pour créer la partie."
 		});
 	}
 
-	// Enregistrer la classe dans la base de données
+	// Vérifie que la classe appartient bien au professeur
+	if(! await isClassBelongsProfessor(req.body.idCourse, req)){
+		return res.status(403).json({
+			message: "Vous n'avez pas accès à cette classe."
+		})
+	}
+
+	// Enregistrer la partie dans la base de données
 	await Game.create({idCourse: req.body.idCourse})
 
 		// Renvoie les données créées
@@ -95,7 +103,7 @@ exports.create = async (req, res) => {
 		// Gère les erreurs
 		.catch(err => {
 			return res.status(500).json({
-				message: err.message || "Une erreur s'est produite lors de la création de la classe."
+				message: err.message || "Une erreur s'est produite lors de la création de la partie."
 			});
 		});
 }
@@ -105,15 +113,49 @@ exports.create = async (req, res) => {
 // 									 READ                                      //
 /////////////////////////////////////////////////////////////////////////////////
 
-// methode pour récuperer une clase du professeur par son id
+// Récupère toutes les parties du professeur connecté
+exports.findAll = async (req, res) => {
+
+	let coursesId = [];
+
+	// Récupère toutes les classes du professeur connecté
+	await Course.findAll({ where: { idProfessor: req.tokenId } })
+	.then(data => {
+		// Récupère les id correspondant aux classes du professeur connécté
+		coursesId = data.map(course => course.dataValues.id);
+	})
+
+	// Gère les erreurs
+	.catch(err => {
+		return res.status(500).json({
+			message: err.message || "Une erreur s'est produite lors de la récupération des classes."
+		});
+	});	
+
+	try {
+		// Récupère toutes les parties correspondantes aux classes du professeur connecté
+		const gamesData = await Game.findAll({ where: { idCourse: { [Op.in]: coursesId } } });
+		return res.status(200).json(gamesData);
+	} catch (err) {
+		return res.status(500).json({
+		message: err.message || "Une erreur s'est produite lors de la récupération des parties."
+		});
+	}
+		  
+	
+
+}
+
+
+// methode pour récuperer une partie en fonction de son id
 exports.findById = async (req, res) => {
 	
 	try{
-		// Vérifie que l'élève appartient bien au professeur
+		// Vérifie que la partie appartient bien au professeur
 		const isBelongsToProfessor = await isGameBelongsProfessor(req.params.id, req);
 		if (!isBelongsToProfessor) {
 			return res.status(403).json({
-				message: "Vous n'avez pas accès à cette classe."
+				message: "Vous n'avez pas accès à cette partie."
 			});
 		}
 
@@ -121,46 +163,60 @@ exports.findById = async (req, res) => {
 			.then(data => {
 				return res.status(200).json(data);
 			})
+
+			// Gère les erreurs
 			.catch(err => {
 				return res.status(500).json({
 					message: err.message || "Une erreur s'est produite lors de la récupération de la partie."
 				});
 			});	
+	// Gère les erreurs
 	} catch (err) {
 		return res.status(500).json({
-			message: err.message || "Une erreur s'est produite lors de la récupération des étudiants."
+			message: err.message || "Une erreur s'est produite lors de la récupération de la partie."
 		});
 	}
 }
-
 
 /////////////////////////////////////////////////////////////////////////////////
 // 									 OTHER                                     //
 /////////////////////////////////////////////////////////////////////////////////
 
 // todo : vérifier si la game appartient au professeur
-// Créer et enregistrer une nouvelle classe au professeurs
+// Ouvre la partie aux élèves
 exports.open = async (req, res) => {
 
+	// Vérifie que la partie appartient bien au professeur
+	const isBelongsToProfessor = await isGameBelongsProfessor(req.params.id, req);
+	if (!isBelongsToProfessor) {
+		return res.status(403).json({
+			message: "Vous n'avez pas accès à cette partie."
+	});
+	}
+
 	GameidCourse = null
+
+	// Récupère l'id de la classe de la partie
 	await Game.findOne({ where: { id: req.params.id} })
 	.then(data => {
 		GameidCourse = data.idCourse;
 	})
+
+	// Gère les erreurs
 	.catch(err => {
 		return res.status(500).json({
 			message: err.message || "Une erreur s'est produite lors de la récupération de la partie."
 		});
 	});	
 
-	// Créer une classe
+	// Créer un code de la partie correspondant aux élèves de la classe concerné par la partie
 	const gameCode = {
 		code: makeid(10),
 		idGame: req.params.id,
 		idCourse: GameidCourse,
 	};
 
-	// Enregistrer la classe dans la base de données
+	// Enregistrer le code dans la base de données
 	await GameCode.create(gameCode)
 
 		// Renvoie les données créées
@@ -176,14 +232,22 @@ exports.open = async (req, res) => {
 		});
 }
 
-
+// todo : le dete renvoie des données ?
+// Ferme la partie aux élèves
 exports.close = async (req, res) => {
 
+	// Vérifie que la partie appartient bien au professeur
+	const isBelongsToProfessor = await isGameBelongsProfessor(req.params.id, req);
+	if (!isBelongsToProfessor) {
+		return res.status(403).json({
+			message: "Vous n'avez pas accès à cette partie."
+	});
+	}
 
 	// Enregistrer la classe dans la base de données
 	await GameCode.destroy({ where: { idGame: req.params.id}})
 
-		// Renvoie les données créées
+		// Renvoie les données supprimées
 		.then(data => {
 			return res.status(201).json(data);
 		})
@@ -196,15 +260,17 @@ exports.close = async (req, res) => {
 		});
 }
 
-// VOBMs6EUub
 
-// methode pour récuperer une clase du professeur par son id
+// methode pour récuperer une classe à partir du code de la partie
 exports.course = async (req, res) => {
-	console.log(req.params.code)
+
+	// Récupère la classe courrespondant au code
 	await GameCode.findOne({ where: { code: req.params.code} })
 		.then(data => {
 			return res.status(200).json(data);
 		})
+
+		// Gère les erreurs
 		.catch(err => {
 			return res.status(500).json({
 				message: err.message || "Une erreur s'est produite lors de la récupération de la partie."
